+++ kubernetes-orchestrator-extension/Jobs/JobBase.cs
@@ -7,6 +7,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
 using System.Linq;
 using System.Security.Cryptography;
@@ -505,16 +506,19 @@ public abstract class JobBase
 
                 break;
             case 2 when IsClusterStore(Capability):
-                Logger.LogWarning("`StorePath`: `{StorePath}` is 2 parts this is not a valid combination for `K8SCluster` and will be ignored", spath);
+                Logger.LogWarning(
+                    "`StorePath`: `{StorePath}` is 2 parts this is not a valid combination for `K8SCluster` and will be ignored",
+                    spath);
                 break;
             case 2 when IsNamespaceStore(Capability):
                 var nsPrefix = sPathParts[0];
                 Logger.LogTrace("nsPrefix: {NsPrefix}", nsPrefix);
                 var nsName = sPathParts[1];
                 Logger.LogTrace("nsName: {NsName}", nsName);
-                
+
                 Logger.LogInformation(
-                    "`StorePath`: `{StorePath}` is 2 parts and store type is `K8SNS`, assuming that store path pattern is either `<cluster_name>/<namespace_name>` or `namespace/<namespace_name>`", spath);
+                    "`StorePath`: `{StorePath}` is 2 parts and store type is `K8SNS`, assuming that store path pattern is either `<cluster_name>/<namespace_name>` or `namespace/<namespace_name>`",
+                    spath);
                 if (string.IsNullOrEmpty(KubeNamespace))
                 {
                     Logger.LogInformation("`KubeNamespace` is empty, setting `KubeNamespace` to `{Namespace}`", nsName);
@@ -522,16 +526,20 @@ public abstract class JobBase
                 }
                 else
                 {
-                    Logger.LogInformation("`KubeNamespace` parameter is not empty, ignoring `StorePath` value `{StorePath}`", spath);
+                    Logger.LogInformation(
+                        "`KubeNamespace` parameter is not empty, ignoring `StorePath` value `{StorePath}`", spath);
                 }
+
                 break;
             case 2:
-                Logger.LogInformation("`StorePath`: `{StorePath}` is 2 parts, assuming that store path pattern is the `<cluster>/<secret_name>` ", spath);
+                Logger.LogInformation(
+                    "`StorePath`: `{StorePath}` is 2 parts, assuming that store path pattern is the `<cluster>/<secret_name>` ",
+                    spath);
                 var kNs = sPathParts[0];
                 Logger.LogTrace("kNs: {KubeNamespace}", kNs);
                 var kSn = sPathParts[1];
                 Logger.LogTrace("kSn: {KubeSecretName}", kSn);
-                
+
                 if (string.IsNullOrEmpty(KubeNamespace))
                 {
                     Logger.LogInformation("`KubeNamespace` is not set, setting `KubeNamespace` to `{Namespace}`", kNs);
@@ -554,10 +562,14 @@ public abstract class JobBase
 
                 break;
             case 3 when IsClusterStore(Capability):
-                Logger.LogError("`StorePath`: `{StorePath}` is 3 parts and store type is `K8SCluster`, this is not a valid combination and `StorePath` will be ignored", spath);
+                Logger.LogError(
+                    "`StorePath`: `{StorePath}` is 3 parts and store type is `K8SCluster`, this is not a valid combination and `StorePath` will be ignored",
+                    spath);
                 break;
             case 3 when IsNamespaceStore(Capability):
-                Logger.LogInformation("`StorePath`: `{StorePath}` is 3 parts and store type is `K8SNS`, assuming that store path pattern is `<cluster>/namespace/<namespace_name>`", spath);
+                Logger.LogInformation(
+                    "`StorePath`: `{StorePath}` is 3 parts and store type is `K8SNS`, assuming that store path pattern is `<cluster>/namespace/<namespace_name>`",
+                    spath);
                 var nsCluster = sPathParts[0];
                 Logger.LogTrace("nsCluster: {NsCluster}", nsCluster);
                 var nsClarifier = sPathParts[1];
@@ -567,33 +579,37 @@ public abstract class JobBase
 
                 if (string.IsNullOrEmpty(KubeNamespace))
                 {
-                    Logger.LogInformation("`KubeNamespace` is not set, setting `KubeNamespace` to `{Namespace}`", nsName3);
+                    Logger.LogInformation("`KubeNamespace` is not set, setting `KubeNamespace` to `{Namespace}`",
+                        nsName3);
                     KubeNamespace = nsName3;
                 }
                 else
                 {
-                    Logger.LogInformation("`KubeNamespace` is set, ignoring `StorePath` value `{StorePath}`", spath);    
+                    Logger.LogInformation("`KubeNamespace` is set, ignoring `StorePath` value `{StorePath}`", spath);
                 }
 
                 if (!string.IsNullOrEmpty(KubeSecretName))
                 {
-                    Logger.LogWarning("`KubeSecretName` parameter is not empty, but is not supported for `K8SNS` store type and will be ignored");
+                    Logger.LogWarning(
+                        "`KubeSecretName` parameter is not empty, but is not supported for `K8SNS` store type and will be ignored");
                     KubeSecretName = "";
                 }
 
                 break;
             case 3:
-                Logger.LogInformation("Store path is 3 parts assuming that it is the '<cluster_name>/<namespace_name>/<secret_name>`");
+                Logger.LogInformation(
+                    "Store path is 3 parts assuming that it is the '<cluster_name>/<namespace_name>/<secret_name>`");
                 var kH = sPathParts[0];
                 Logger.LogTrace("kH: {KubeHost}", kH);
                 var kN = sPathParts[1];
                 Logger.LogTrace("kN: {KubeNamespace}", kN);
                 var kS = sPathParts[2];
                 Logger.LogTrace("kS: {KubeSecretName}", kS);
-                
+
                 if (kN is "secret" or "tls" or "certificate" or "namespace")
                 {
-                    Logger.LogInformation("Store path is 3 parts and the second part is a reserved keyword, assuming that it is the '<cluster_name>/<namespace_name>/<secret_name>'");
+                    Logger.LogInformation(
+                        "Store path is 3 parts and the second part is a reserved keyword, assuming that it is the '<cluster_name>/<namespace_name>/<secret_name>'");
                     kN = sPathParts[0];
                     kS = sPathParts[1];
                 }
@@ -826,6 +842,14 @@ public abstract class JobBase
             KubeSvcCreds = ServerPassword;
         }
 
+        if (string.IsNullOrEmpty(KubeSvcCreds))
+        {
+            const string credsErr =
+                "No credentials provided to connect to Kubernetes. Please provide a kubeconfig file. See https://github.com/Keyfactor/kubernetes-orchestrator/blob/main/scripts/kubernetes/get_service_account_creds.sh";
+            Logger.LogError(credsErr);
+            throw new ConfigurationException(credsErr);
+        }
+
         switch (KubeSecretType)
         {
             case "pfx":
@@ -877,15 +901,15 @@ public abstract class JobBase
         }
 
         Logger.LogTrace("Creating new KubeCertificateManagerClient object");
-        // KubeClient = new KubeCertificateManagerClient(KubeSvcCreds);
-        //
-        // Logger.LogTrace("Getting KubeHost and KubeCluster from KubeClient");
-        // KubeHost = KubeClient.GetHost();
-        // Logger.LogTrace("KubeHost: {KubeHost}", KubeHost);
-        //
-        // Logger.LogTrace("Getting cluster name from KubeClient");
-        // KubeCluster = KubeClient.GetClusterName();
-        // Logger.LogTrace("KubeCluster: {KubeCluster}", KubeCluster);
+        KubeClient = new KubeCertificateManagerClient(KubeSvcCreds);
+
+        Logger.LogTrace("Getting KubeHost and KubeCluster from KubeClient");
+        KubeHost = KubeClient.GetHost();
+        Logger.LogTrace("KubeHost: {KubeHost}", KubeHost);
+
+        Logger.LogTrace("Getting cluster name from KubeClient");
+        KubeCluster = KubeClient.GetClusterName();
+        Logger.LogTrace("KubeCluster: {KubeCluster}", KubeCluster);
 
         if (string.IsNullOrEmpty(KubeSecretName) && !string.IsNullOrEmpty(StorePath) && !Capability.Contains("NS") &&
             !Capability.Contains("Cluster"))
@@ -931,17 +955,11 @@ public abstract class JobBase
 
 
             if (Capability.Contains("K8SNS"))
-            {
                 secretType = "namespace";
-            }
             else if (Capability.Contains("K8SCluster"))
-            {
                 secretType = "cluster";
-            }
             else
-            {
                 secretType = KubeSecretType.ToLower();
-            }
 
             Logger.LogTrace("secretType: {SecretType}", secretType);
             Logger.LogTrace("Entered switch statement based on secretType");
@@ -990,6 +1008,23 @@ public abstract class JobBase
         {
             Logger.LogError("Unknown error constructing canonical store path {Error}", e.Message);
             return StorePath;
+        }
+    }
+
+    protected string ResolvePamField(string name, string value)
+    {
+        try
+        {
+            Logger.LogTrace($"Attempting to resolved PAM eligible field {name}");
+            return _resolver.Resolve(value);
+        }
+        catch (Exception e)
+        {
+            Logger.LogError($"Unable to resolve PAM field {name}. Returning original value.");
+            Logger.LogError(e.Message);
+            Logger.LogTrace(e.ToString());
+            Logger.LogTrace(e.StackTrace);
+            return value;
         }
     }
 
@@ -1047,7 +1082,6 @@ public abstract class JobBase
             try
             {
                 if (certObj.HasPrivateKey)
-                {
                     try
                     {
                         Logger.LogDebug("Attempting to export private key as PKCS8");
@@ -1076,7 +1110,6 @@ public abstract class JobBase
                         Logger.LogTrace("ExportEncryptedPkcs8PrivateKey() complete");
                         return keyBytes;
                     }
-                }
             }
             catch (Exception ie)
             {
@@ -1105,13 +1138,10 @@ public abstract class JobBase
         var result = new JobResult
         {
             Result = OrchestratorJobStatusJobResult.Success,
-            JobHistoryId = jobHistoryId,
+            JobHistoryId = jobHistoryId
         };
 
-        if (!string.IsNullOrEmpty(jobMessage))
-        {
-            result.FailureMessage = jobMessage;
-        }
+        if (!string.IsNullOrEmpty(jobMessage)) result.FailureMessage = jobMessage;
 
         return result;
     }
@@ -1238,15 +1268,11 @@ public abstract class JobBase
             Logger.LogDebug("No password found");
             var passwdEx = "";
             if (!string.IsNullOrEmpty(StorePasswordPath))
-            {
                 passwdEx = "Store secret '" + StorePasswordPath + "'did not contain key '" + CertificateDataFieldName +
                            "' or '" + PasswordFieldName + "'" +
                            "  Please provide a valid store password and try again";
-            }
             else
-            {
                 passwdEx = "Invalid store password.  Please provide a valid store password and try again";
-            }
 
             Logger.LogError("{Msg}", passwdEx);
             throw new Exception(passwdEx);
@@ -1278,10 +1304,7 @@ public abstract class JobBase
     protected string GetCertificatePem(Pkcs12Store store, string password, string alias = "")
     {
         Logger.LogDebug("Entered GetCertificatePem()");
-        if (string.IsNullOrEmpty(alias))
-        {
-            alias = store.Aliases.Cast<string>().FirstOrDefault(store.IsKeyEntry);
-        }
+        if (string.IsNullOrEmpty(alias)) alias = store.Aliases.Cast<string>().FirstOrDefault(store.IsKeyEntry);
 
         Logger.LogDebug("Attempting to get certificate with alias {Alias}", alias);
         var cert = store.GetCertificate(alias).Certificate;